pyproject.toml
[project]
name = "mcp-client-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.10.1",
    "pyyaml>=6.0.2",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.2",
    "httpx>=0.25.0",
]
